# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CloudPAM API
  version: "0.7.0"
  description: |
    REST API for managing CloudPAM pools, accounts, blocks, data exports,
    authentication, audit logging, schema planning, cloud resource discovery,
    and actionable recommendations.
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/test-sentry:
    get:
      summary: Test Sentry integration
      tags: [System]
      security: []
      parameters:
        - name: type
          in: query
          description: "Test type: message, error, or panic"
          schema:
            type: string
            enum: [message, error, panic]
      responses:
        "200":
          description: Test executed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  usage:
                    type: string
  /api/v1/pools:
    get:
      summary: List pools
      tags: [Pools]
      parameters:
        - name: include_stats
          in: query
          description: Include utilization statistics for each pool
          schema:
            type: boolean
      responses:
        "200":
          description: Pools returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pool"
        default:
          $ref: "#/components/responses/Error"
    post:
      summary: Create pool
      tags: [Pools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePool"
      responses:
        "201":
          description: Pool created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pool"
        default:
          $ref: "#/components/responses/Error"
...
  /api/v1/accounts:
    get:
      summary: List accounts
      tags: [Accounts]
      responses:
        "200":
          description: Accounts returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
        default:
          $ref: "#/components/responses/Error"
    post:
      summary: Create account
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccount"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        default:
          $ref: "#/components/responses/Error"
...
  /api/v1/blocks:
    get:
      summary: List assigned sub-pools
      tags: [Blocks]
      parameters:
        - $ref: "#/components/parameters/AccountsFilter"
        - $ref: "#/components/parameters/PoolsFilter"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Assigned blocks listed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignedBlocksResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/export:
    get:
      summary: Export datasets as CSV inside ZIP archive
      tags: [Export]
      parameters:
        - name: datasets
          in: query
          description: Comma-separated datasets to include (accounts, pools, blocks)
          required: true
          schema:
            type: string
        - name: accounts_fields
          in: query
          description: Comma-separated account columns to include
          schema:
            type: string
        - name: pools_fields
          in: query
          description: Comma-separated pool columns to include
          schema:
            type: string
        - name: blocks_fields
          in: query
          description: Comma-separated block columns to include
          schema:
            type: string
        - $ref: "#/components/parameters/AccountsFilter"
        - $ref: "#/components/parameters/PoolsFilter"
      responses:
        "200":
          description: ZIP archive streaming response
          content:
            application/zip:
              schema:
                type: string
                format: binary
        default:
          $ref: "#/components/responses/Error"
  /api/v1/import/accounts:
    post:
      summary: Import accounts from CSV
      tags: [Import]
      requestBody:
        required: true
        content:
          text/csv:
            schema:
              type: string
              description: CSV with header row. Required columns; key, name. Optional; provider, external_id, description, platform, tier, environment, regions.
      responses:
        "200":
          description: Import results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/import/pools:
    post:
      summary: Import pools from CSV
      tags: [Import]
      requestBody:
        required: true
        content:
          text/csv:
            schema:
              type: string
              description: CSV with header row. Required columns; name, cidr. Optional; parent_id, account_id, account_key, type, status, source, description.
      responses:
        "200":
          description: Import results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/schema/check:
    post:
      summary: Check schema for conflicts with existing pools
      tags: [Schema]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchemaCheckRequest"
      responses:
        "200":
          description: Conflict check results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaCheckResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/schema/apply:
    post:
      summary: Bulk-create pools from a schema plan
      tags: [Schema]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchemaApplyRequest"
      responses:
        "200":
          description: Schema apply results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaApplyResponse"
        "409":
          description: Conflict with existing pools (when skip_conflicts is false)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/search:
    get:
      summary: Search across pools and accounts with CIDR containment queries
      tags: [Search]
      parameters:
        - name: q
          in: query
          description: Free-text search matching name, CIDR, key, or description
          schema:
            type: string
        - name: cidr_contains
          in: query
          description: Find pools whose CIDR contains this IP or prefix (e.g., "10.1.2.5" or "10.1.0.0/16")
          schema:
            type: string
        - name: cidr_within
          in: query
          description: Find pools within this prefix (e.g., "10.0.0.0/8")
          schema:
            type: string
        - name: type
          in: query
          description: "Comma-separated entity types to search: pool, account"
          schema:
            type: string
            example: "pool,account"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          description: Invalid search parameters (e.g., bad CIDR format)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API key authentication. Keys use the format `cpam_<base64url>`.
        Pass in the Authorization header as `Bearer cpam_...`.
  parameters:
...
    Page:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
    PageSize:
      name: page_size
      in: query
      description: Number of items per page; specify "all" to disable pagination
      schema:
        oneOf:
          - type: integer
            minimum: 0
          - type: string
            enum: ["all"]
    AccountsFilter:
      name: accounts
      in: query
      description: Comma-separated account IDs to filter by
      schema:
        type: string
        example: "1,2"
    PoolsFilter:
      name: pools
      in: query
      description: Comma-separated parent pool IDs to filter by
      schema:
        type: string
        example: "10,11"
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
      required: [error]
    PoolType:
      type: string
      enum: [supernet, region, environment, vpc, subnet]
    PoolStatus:
      type: string
      enum: [planned, active, deprecated]
    PoolSource:
      type: string
      enum: [manual, discovered, imported]
    Pool:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        cidr:
          type: string
        parent_id:
          type: integer
          format: int64
          nullable: true
        account_id:
          type: integer
          format: int64
          nullable: true
        type:
          $ref: "#/components/schemas/PoolType"
        status:
          $ref: "#/components/schemas/PoolStatus"
        source:
          $ref: "#/components/schemas/PoolSource"
        description:
          type: string
        tags:
          type: object
          additionalProperties:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, cidr, type, status, source, created_at, updated_at]
...
    CreatePool:
      type: object
      properties:
        name:
          type: string
        cidr:
          type: string
        parent_id:
          type: integer
          format: int64
          nullable: true
        account_id:
          type: integer
          format: int64
          nullable: true
        type:
          $ref: "#/components/schemas/PoolType"
        status:
          $ref: "#/components/schemas/PoolStatus"
        source:
          $ref: "#/components/schemas/PoolSource"
        description:
          type: string
        tags:
          type: object
          additionalProperties:
            type: string
      required: [name, cidr]
...
    Account:
      type: object
      properties:
        id:
          type: integer
          format: int64
        key:
          type: string
        name:
          type: string
        provider:
          type: string
        external_id:
          type: string
        description:
          type: string
        platform:
          type: string
        tier:
          type: string
        environment:
          type: string
        regions:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
      required: [id, key, name, created_at]
    CreateAccount:
      type: object
      properties:
        key:
          type: string
        name:
          type: string
        provider:
          type: string
        external_id:
          type: string
        description:
          type: string
        platform:
          type: string
        tier:
          type: string
        environment:
          type: string
        regions:
          type: array
          items:
            type: string
      required: [key, name]
...
    AssignedBlock:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        cidr:
          type: string
        parent_id:
          type: integer
          format: int64
        parent_name:
          type: string
        account_id:
          type: integer
          format: int64
          nullable: true
        account_name:
          type: string
        account_platform:
          type: string
        account_tier:
          type: string
        account_environment:
          type: string
        account_regions:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
      required: [id, name, cidr, parent_id, parent_name, created_at]
    AssignedBlocksResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AssignedBlock"
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
      required: [items, total, page, page_size]
    ImportResult:
      type: object
      properties:
        created:
          type: integer
          description: Number of records successfully created
        skipped:
          type: integer
          description: Number of records skipped (duplicates)
        errors:
          type: array
          items:
            type: string
          description: List of row-level error messages
      required: [created, skipped, errors]
...
    SchemaPoolEntry:
      type: object
      properties:
        ref:
          type: string
          description: Unique reference ID within this schema
        name:
          type: string
        cidr:
          type: string
        type:
          type: string
          enum: [supernet, region, environment, vpc, subnet]
        parent_ref:
          type: string
          description: Reference to parent pool's ref (empty for root)
      required: [ref, name, cidr, type]
    SchemaCheckRequest:
      type: object
      properties:
        pools:
          type: array
          items:
            $ref: "#/components/schemas/SchemaPoolEntry"
      required: [pools]
    SchemaConflict:
      type: object
      properties:
        planned_cidr:
          type: string
        planned_name:
          type: string
        existing_pool_id:
          type: integer
          format: int64
        existing_pool_name:
          type: string
        existing_cidr:
          type: string
        overlap_type:
          type: string
          enum: [overlap, contains, contained_by]
      required: [planned_cidr, planned_name, existing_pool_id, existing_pool_name, existing_cidr, overlap_type]
    SchemaCheckResponse:
      type: object
      properties:
        conflicts:
          type: array
          items:
            $ref: "#/components/schemas/SchemaConflict"
        total_pools:
          type: integer
        conflict_count:
          type: integer
      required: [conflicts, total_pools, conflict_count]
    SchemaApplyRequest:
      type: object
      properties:
        pools:
          type: array
          items:
            $ref: "#/components/schemas/SchemaPoolEntry"
          description: Pools in topological order (parents before children)
        status:
          type: string
          enum: [planned, active, deprecated]
          description: Status to assign to all created pools (default "planned")
        tags:
          type: object
          additionalProperties:
            type: string
          description: Tags to apply to all created pools
        skip_conflicts:
          type: boolean
          description: Skip conflict check with existing pools (default false)
      required: [pools]
    SchemaApplyResponse:
      type: object
      properties:
        created:
          type: integer
          description: Number of pools successfully created
        skipped:
          type: integer
          description: Number of pools skipped due to errors
        errors:
          type: array
          items:
            type: string
          description: List of error messages for failed pools
        root_pool_id:
          type: integer
          format: int64
          description: ID of the root pool (first pool without a parent_ref)
        pool_map:
          type: object
          additionalProperties:
            type: integer
            format: int64
          description: Mapping of ref to created pool ID
      required: [created, skipped, errors, root_pool_id, pool_map]
...
    SearchResultItem:
      type: object
      properties:
        type:
          type: string
          enum: [pool, account]
        id:
          type: integer
          format: int64
        name:
          type: string
        cidr:
          type: string
          description: Pool CIDR (only for pool results)
        description:
          type: string
        status:
          type: string
          description: Pool status (only for pool results)
        pool_type:
          type: string
          description: Pool type (only for pool results)
        account_key:
          type: string
          description: Account key (only for account results)
        provider:
          type: string
          description: Account provider (only for account results)
        parent_id:
          type: integer
          format: int64
          nullable: true
        account_id:
          type: integer
          format: int64
          nullable: true
      required: [type, id, name]
    SearchResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SearchResultItem"
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
      required: [items, total, page, page_size]
...